# docs/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stratum — organizations, accounts, billing, and notifications.",
        "title": "Stratum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/organizations/{organizationID}/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cursor-paginated file listing, optionally scoped to a folder or filtered by search. Admin/owner only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "List files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "filter by folder",
                        "name": "folder_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "search across all folders instead of just folder_id",
                        "name": "search_all",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagination cursor from a previous response",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/fileRecord"
                                                    }
                                                },
                                                "next_cursor": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin/owner only. Max 50MB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination folder ID (root if omitted)",
                        "name": "folder_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/fileRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "missing file or too large",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationID}/files/bulk-delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin/owner only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Bulk soft-delete files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File IDs to delete",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bulkDeleteFilesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "deleted_count": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
        "/organizations/{organizationID}/files/trash": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns up to the 50 most recently soft-deleted files. Admin/owner only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "List deleted files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/fileRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
        "/organizations/{organizationID}/files/{fileID}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Redirects to a short-lived signed download URL. Admin/owner only.",
                "tags": [
                    "organization"
                ],
                "summary": "Download a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "redirect to signed URL"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "404": {
                        "description": "file not found",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Payload": {
            "type": "object",
            "properties": {
                "data": {},
                "next_cursor": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/pagination"
                },
                "status": {
                    "$ref": "#/definitions/Status"
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "bulkDeleteFilesRequest": {
            "type": "object",
            "required": [
                "file_ids"
            ],
            "properties": {
                "file_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "fileRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                }
            }
        },
...
        "pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token issued by Supabase auth.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}